+++ tests/unit/specs/components/transactions/TransactionMetadata.spec.js
@@ -8,14 +8,22 @@ describe(`TransactionMetadata`, () => {
     wrapper = shallowMount(TransactionMetadata, {
       propsData: {
         transaction: {
-          timestamp: new Date("2019-07-31"),
-          height: 1234567,
-          fee: {
-            amount: "0.001234",
-            denom: "ATOM"
-          },
-          hash: `ABCD1234`,
-          undelegationEndTime: `2019-11-16T14:08:20Z`
+          type: "SendTx",
+          hash:
+            "A0DEB29E97A4DF38289D55D63C5724588985E1D35B26518CB66EAF96CFEF2E04",
+          height: 123,
+          timestamp: "2020-02-10T10:15:51Z",
+          memo: "(Sent via Lunie)",
+          success: true,
+          fees: [],
+          details: {
+            from: ["cosmos123"],
+            to: ["cosmos456"],
+            amount: {
+              denom: "ATOM",
+              amount: "1"
+            }
+          }
         }
       },
       stubs: [`router-link`]
